handle isFavorite

+++ feature/newsList/src/main/java/com/example/newslist/ui/NewsViewModel.kt
@@ -2,8 +2,6 @@ package com.example.newslist.ui
 
 
 import android.util.Log
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.core.Result
@@ -11,8 +9,8 @@ import com.example.core.models.Article
 import com.example.core.models.NewsResponse
 import com.example.core.repository.NewsRepo
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -20,16 +18,18 @@ import javax.inject.Inject
 class NewsViewModel @Inject constructor(
     private val newsRepository: NewsRepo
 ) : ViewModel() {
-    private val _newsListFlow: MutableStateFlow<NewsUiState> =
-        MutableStateFlow(NewsUiState())
-    val newsListFlow: Flow<NewsUiState> = _newsListFlow
+    private val _newsListFlow: MutableStateFlow<NewsUiState> = MutableStateFlow(NewsUiState())
+    val newsListFlow: StateFlow<NewsUiState> = _newsListFlow
     private var breakingNewsPage = 1
 
     init {
-        getBreakingNews()
+        viewModelScope.launch {
+            getSavedNews()
+            getBreakingNews()
+        }
     }
 
-    private fun getBreakingNews() = viewModelScope.launch {
+    private suspend fun getBreakingNews() {
         _newsListFlow.value = _newsListFlow.value.copy(isLoading = true)
         val response = newsRepository.getBreakingNews("us", breakingNewsPage)
         Log.d("loadMore", "getBreakingNews: $breakingNewsPage")
@@ -37,11 +37,15 @@ class NewsViewModel @Inject constructor(
     }
 
     fun loadMore() {
-        getBreakingNews()
+        viewModelScope.launch {
+            getBreakingNews()
+        }
     }
 
     fun retry() {
-        getBreakingNews()
+        viewModelScope.launch {
+            getBreakingNews()
+        }
     }
 
 
@@ -50,14 +54,13 @@ class NewsViewModel @Inject constructor(
             is Result.Success -> {
                 response.data.let { resultResponse ->
                     breakingNewsPage++
-                    _newsListFlow.value = _newsListFlow.value.copy(
-                        isLoading = false,
+                    _newsListFlow.value = _newsListFlow.value.copy(isLoading = false,
                         error = null,
-                        articles =
-                        mutableListOf<Article>().apply {
-                            addAll(_newsListFlow.value.articles + resultResponse.articles)
-                        }
-                    )
+                        articles = mutableListOf<ArticleUiState>().apply {
+                            addAll(_newsListFlow.value.articles + resultResponse.articles.map {
+                                ArticleUiState(it, savedNews.map { it.url }.contains(it.url))
+                            })
+                        })
                 }
             }
 
@@ -69,23 +72,25 @@ class NewsViewModel @Inject constructor(
 
     }
 
-    /* fun saveArticle(article: Article) = viewModelScope.launch {
-         newsRepository.upsert(article)
-         getSavedNews()
-     }
+    private fun refreshNews() {
+        val refreshedState = _newsListFlow.value.copy(articles = _newsListFlow.value.articles.map {
+            ArticleUiState(it.article, savedNews.map { it.url }.contains(it.article.url))
+        })
+        _newsListFlow.value = refreshedState
+    }
 
-     private val _savedNews = MutableLiveData<List<Article>>()
-     val savedNews: LiveData<List<Article>> = _savedNews
-     fun getSavedNews() {
-         viewModelScope.launch {
-             _savedNews.postValue(newsRepository.getSavedNews())
-         }
-     }
+    private var savedNews: Set<Article> = setOf()
 
-     fun deleteArticle(article: Article) = viewModelScope.launch {
-         newsRepository.deleteArticle(article)
-         getSavedNews()
-     }*/
+    private suspend fun getSavedNews() {
+        savedNews = newsRepository.getSavedNews().toSet()
+    }
+
+    fun addRemoveArticle(article: Article) = viewModelScope.launch {
+        if (savedNews.map { it.url }.contains(article.url)) newsRepository.deleteArticle(article)
+        else newsRepository.upsert(article)
+        getSavedNews()
+        refreshNews()
+    }
 }
